+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Token Preview not updating when changing the token image path.
+- Removing Predicates not returning to default value of `[]`.
 
 ## [0.1.2] - 2024-08-02